requirements.txt
# Data Processing
pandas>=1.5.0
numpy>=1.21.0
pyspark>=3.3.0  # Optional, for large-scale data processing

# Configuration Management
pyyaml>=6.0.0

# Data Storage
pyarrow>=10.0.0  # Parquet support

# Development Tools
pytest>=7.0.0  # Testing framework
black>=22.0.0  # Code formatting
isort>=5.10.0  # Import sorting
mypy>=0.950  # Type checking 